scrapper_backend/requirements.txt
Django==4.0.2
psycopg2==2.9.3
djangorestframework==3.13.1